Suppress the final set of warnings in pcl apps

+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/pipeline/impl/global_nn_recognizer_cvfh.hpp
@@ -295,7 +295,7 @@ template<template<class > class Distance, typename PointInT, typename FeatureT>
               nearestKSearch (single_categories_index_[it->second], histogram, NN_, indices, distances);
               //gather NN-search results
               double score = 0;
-              for (size_t i = 0; i < NN_; ++i)
+              for (size_t i = 0; i < (size_t) NN_; ++i)
               {
                 score = distances[0][i];
                 index_score is;
